Fix: appendToBody with jQuery < 1.10.0; stealing focus #944
.relpaceWith in jQuery (version less then 1.10) doesn't work properly
and as so, ui-select with 'appendToBody=true' doesn't display selected
value.
Use .after() and .remove() instead of .replaceWith() to avoid this
issue.

fix for #944: Don't restore focus if it was moved out of select (irregardles of target
element). Otherwise it looks too confusing when clicking on links or any
other control that is not an input/button/textarea.